// interfaces/enums/Ticker.ts
enum Ticker {
  None = 'NONE',

  // Regular Tokens
  ETH = 'ETH',
  USDT = 'USDT',
  USDC = 'USDC',
  SUSHI = 'SUSHI',
  DAI = 'DAI',
  YFI = 'YFI',
  LINK = 'LINK',
  SUSD = 'SUSD',
  LEND = 'LEND',
  AMPL = 'AMPL',
  WBTC = 'WBTC',
  UNI = 'UNI',
  SWSH = 'SWSH',
  RTC = 'RTC',
  LIQLO = 'LIQLO',
  SPECTRE = 'SPECTRE',
  PSHP = 'PSHP',

  // Liquidity Tokens
  USDT_ETH_SP = 'USDT-ETH_SP',
  USDC_ETH_SP = 'USDC-ETH_SP',
  SUSHI_ETH_SP = 'SUSHI-ETH_SP',
  DAI_ETH_SP = 'DAI-ETH_SP',
  YFI_ETH_SP = 'YFI-ETH_SP',
  LINK_ETH_SP = 'LINK-ETH_SP',
  SUSD_ETH_SP = 'SUSD-ETH_SP',
  LEND_ETH_SP = 'LEND-ETH_SP',
  DAI_ETH_UNIV2 = 'DAI-ETH_UNIV2',
  ETH_AMPL_UNIV2 = 'ETH-AMPL_UNIV2',
  SWSH_ETH_UNIV2 = 'SWSH-ETH_UNIV2',
  RTC_ETH_UNIV2 = 'RTC-ETH_UNIV2',
  LIQLO_ETH_UNIV2 = 'LIQLO-ETH_UNIV2',
  LIQLO_RTC_UNIV2 = 'LIQLO-RTC_UNIV2',
  SWSH_RTC_UNIV2 = 'SWSH-RTC_UNIV2',
  SWSH_SPECTRE_UNIV2 = 'SWSH-SPECTRE_UNIV2',
  SPECTRE_RTC_UNIV2 = 'SPECTRE-RTC_UNIV2',
  SPECTRE_ETH_UNIV2 = 'SPECTRE-ETH_UNIV2',
  WBTC_ETH_UNIV2 = 'WBTC-ETH_UNIV2',
  USDT_ETH_UNIV2 = 'USDT-ETH_UNIV2',
  USDC_ETH_UNIV2 = 'USDC-ETH_UNIV2',
  UNI_ETH_UNIV2 = 'UNI-ETH_UNIV2',
  PSHP_ETH_UNIV2 = 'PSHP-ETH_UNIV2'
}

export default Ticker